Bootstrapping and configuration code for reversed row_key_split_index table
We wish to reverse the order of the index on our reverse lookup table, however this is not possible to do in place on our existing dataset.  Instead, a parallel index will be filled alongside the current one, which will be used for queries whose range falls entirely within its contents, falling back to the old index for others.

Two new flags will be introduced to control this function:
```
kairosdb.datastore.cassandra.use-new-split-index=false
kairosdb.datastore.cassandra.new-split-index-start-time-ms=0l
```

The new-split-index-start-time-ms value will be set to some value after the new index begins to fill, and the check for which index to use will be startTime > new-split-index-start-time-ms.

Updates to the read/write paths will follow in subsequent commits.