Retrieve a QueryMetric only once within a query
Make use of the internal caching mechanism to ensure that a given
QueryMetric query (which results in a query in the underlying
datastore) is only performed once within the context of a single
query operation in the REST interface.

This change can greatly improve performance for the case in which
multiple groupings are being requested for the same range of data
(for example, min, max, avg for a given range). It is also an
improvement in terms of correctness, as retrieving multiple groupings
of the same range of data will now use a stable view of the queried
data.

This change works regardless of the caching settings that are
provided within the query.