Make the DataGapsMarkingAggregator end-inclusive
Make the behavior of the DataGapsMarkingAggregator consistent
regardless of whether there is data on the "until" endpoint of
a query or not -- the sample point starting at the end poing
is always included. This is a change in the RangeAggregator, and
brings it in line with the general concept of the end range
being inclusive everywhere.

This commit also adds an initial DataPointMatcher Hamcrest matcher
to make matching DataPoints in tests easier.