Use JDK classes for DataCache
Minor change to use built-in JDK functionality for the DataCache
class. This is just an opportunity to remove some code, as well as
a minor speed increase because it only requires calculating the
hashCode for stored elements once (instead of once during a store
operation and once during eviction of the oldest element).