Retry async write failures
Retry writes that fail on async writes to the string_index or
row_keys tables, so that failure to write to these tables will
result in the full batch being retried.

In practice, writes to these tables are quite low in frequency,
but their correct population (particularly the row_keys table)
is essential for correctness, and a failed write will be cached
as having already been written and so won't be retried.

This commit verifies that these async writes were successful
and fails the full batch if they are not successful (thereby
triggering a retry).