eagleeye/ZMON#270 move calculated Xmx/Xms before JAVA_OPTS
This allows us to override this calculation in the deployment settings
The reason behind this is that in docker/kubernetes environment using
`/proc/meminfo` will give the whole amount of memory if the host, not
the limit of the docker. Which leads to the quite big heap, without any
garbage collection, and once we reach the limit which is assigned to the
container, the application is OOM killed.