makes rollup sample periods deterministic
If the rollup sample size is multiple units, align the first sample start time
on the boundary of the next highest granularity. Eg a 10 min sample size should
start on minute zero of hour. Eg a quarterly rollup (3 MONTHS) should start on
the first of January, or the first of April, etc.

This also makes the RollupProcessorImpl test deterministic and it won't fail
depending on the current clock minute, as the current clock minute was the
sample boundary.